Fixed Lua errors stopping scripts from working
If an errors occurs the thread becomes unusable, we have to create a new
one in that case. Additionally, every thread that is created is added to
the global stack, which slowly fills up. The thread has to be removed
after we're done with the NPC. Every time you talk to an NPC now a new
thread is created, for the duration of the conversation and removed
afterwards or on error.